Refactor GraphQLDocumentParser
- Extract exceptions
- Extract some utils extensions
- Extract pos parse validations
